refactor: extract shared ESM resolution, handle nested exports, fix shadowing
- Extract resolveEsmEntry() into shared esm-resolve.ts module (DRY)
- Handle nested conditional exports where exports["."].import is an
  object with a "default" key (e.g. { types: "...", default: "..." })
- Hoist airJsonDir declaration in prepare.ts to eliminate variable shadowing
